requirements.txt
requests==2.28.2
mysql-connector-python
python-dotenv